SVG: Move declaration of templated methods to inline header to fix non-unified builds https://bugs.webkit.org/show_bug.cgi?id=296178
Reviewed by Rob Buis.

With non-unified builds, the implementation of both
parseAndReplaceTransform() and parseTransform() don't get
instantiated for all the template types they are used with,
which results in a linking error. This can be fixed by moving
the implementation of those methods to an inline header file that
can then be included where appropriate, so that the methods are
instantiated in those compilation units as expected.

This linking error doesn't happen with unified builds, as
all the related SVG files get bundled into one compilation unit.

* Source/WebCore/svg/SVGTransformList.cpp:
* Source/WebCore/svg/SVGTransformable.cpp:
(WebCore::parseTransformParamList): Deleted.
(WebCore::SVGTransformable::parseAndReplaceTransform): Deleted.
(WebCore::SVGTransformable::parseTransform): Deleted.
* Source/WebCore/svg/SVGTransformable.h:
* Source/WebCore/svg/SVGTransformableInlines.h: Copied from Source/WebCore/svg/SVGTransformable.cpp.
(WebCore::parseTransformParamList):
(WebCore::SVGTransformable::parseAndReplaceTransform):
(WebCore::SVGTransformable::parseTransform):

Canonical link: https://commits.webkit.org/297595@main
