+++ Source/WebCore/svg/SVGTransformList.cpp
@@ -30,6 +30,7 @@
 #include "SVGParserUtilities.h"
 #include "SVGTransform.h"
 #include "SVGTransformable.h"
+#include "SVGTransformableInlines.h"
 #include <wtf/text/StringBuilder.h>
 #include <wtf/text/StringParsingBuffer.h>
